Add PayPalExperienceProfile as a shared object to be use for Checkout and Vault encodable objects
